% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{delim2h5}
\alias{delim2h5}
\title{Convert to HDF5 with a custom callback}
\usage{
delim2h5(
  input_files,
  output_file,
  h5_args = list(datapath = "/"),
  callback_fun = write_df_h5,
  id_col = NULL,
  ...
)
}
\arguments{
\item{output_file}{output HDF5 file}

\item{h5_args}{... args for write_df_h5 unpacked and passed to `callback_fun`}

\item{callback_fun}{function with signature matching function(df,filename,datapath,...) (defaults to `write_df_h5`)}

\item{...}{}

\item{input_file}{one or more files able to be read by `readr::read_delim`}
}
\value{

}
\description{
Convert to HDF5 with a custom callback
}
\examples{

temp_h5 <- fs::file_temp(ext="h5")
delim2h5(readr::readr_example("mtcars.csv"),temp_h5,delim="/")
new_data  <- read_df_h5(temp_h5)
}
