\name{EigenH5-package}
\alias{EigenH5-package}
\alias{EigenH5}
\docType{package}
\title{
\packageTitle{EigenH5}
}
\description{
\packageDescription{EigenH5}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EigenH5}
\packageIndices{EigenH5}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{EigenH5}

Maintainer: \packageMaintainer{EigenH5}
}
\references{
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
