% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rle2offset}
\alias{rle2offset}
\title{Convert RLE-encoded vector to offset+size dataframe}
\usage{
rle2offset(x, na_replace = -1L)
}
\arguments{
\item{x}{either a vector of class `rle`, or a vector that can be converte to one via (`rle(x)`)}

\item{na_replace}{value to replace NA (rle doesn't play well with NA)}
}
\value{
tibble with columns `value`,`offset` and `datasize`
}
\description{
Convert RLE-encoded vector to offset+size dataframe
}
\examples{
x <- rev(rep(6:10, 1:5))
x_na <- c(NA,NA,NA,rev(rep(6:10, 1:5)))
x_n3 <- c(-3,-3,-3,rev(rep(6:10, 1:5)))
print(rle2offset(rle(x)))
stopifnot(
  identical(rle2offset(rle(x)),rle2offset(x)),
  identical(rle2offset(x_na,na_replace=-3),rle2offset(x_n3)))
}
